# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachIntakeConversationRequest": {
        "additionalProperties": false,
        "description": "Accept only a bounded provider conversation identifier from the SDK.",
        "properties": {
          "conversation_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Conversation Id",
            "type": "string"
          }
        },
        "required": [
          "conversation_id"
        ],
        "title": "AttachIntakeConversationRequest",
        "type": "object"
      },
...
      "CreateIntakeSessionRequest": {
        "additionalProperties": false,
        "description": "Select the privacy boundary before creating a browser voice session.",
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/IntakeDataMode",
            "default": "supervised_role_play"
          }
        },
        "title": "CreateIntakeSessionRequest",
        "type": "object"
      },
...
      "DocumentIntakeRequest": {
        "additionalProperties": false,
        "description": "Bound the unstructured document text accepted by the intake route.",
        "properties": {
          "document_text": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Document Text",
            "type": "string"
          }
        },
        "required": [
          "document_text"
        ],
        "title": "DocumentIntakeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "WebhookAck": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          }
        },
        "required": [
          "accepted",
          "duplicate"
        ],
        "title": "WebhookAck",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Mock-first moving-services negotiation API",
    "title": "VeraMove API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/calls/{call_id}/recording": {
      "get": {
        "operationId": "get_call_recording_api_calls__call_id__recording_get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Call Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "signature",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{64}$",
              "title": "Signature",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "audio/mp4": {},
              "audio/mpeg": {},
              "audio/wav": {}
            },
            "description": "Validated provider recording audio."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Call Recording",
        "tags": [
          "calls"
        ]
      }
    },
    "/api/calls/{call_id}/repair": {
      "post": {
        "operationId": "repair_call_api_calls__call_id__repair_post",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Call Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-veramove-operator-secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Veramove-Operator-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repair Call",
        "tags": [
          "calls"
        ]
      }
    },
    "/api/intake/conversations/{conversation_id}": {
      "get": {
        "operationId": "get_intake_session_by_conversation_api_intake_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intake Session By Conversation",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/document": {
      "post": {
        "operationId": "create_job_from_document_api_intake_document_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job From Document",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions": {
      "post": {
        "operationId": "create_intake_session_api_intake_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CreateIntakeSessionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Intake Session",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}": {
      "get": {
        "operationId": "get_intake_session_api_intake_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intake Session",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}/conversation": {
      "post": {
        "operationId": "attach_browser_voice_conversation_api_intake_sessions__session_id__conversation_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachIntakeConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Browser Voice Conversation",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}/finish-manually": {
      "post": {
        "operationId": "finish_intake_manually_api_intake_sessions__session_id__finish_manually_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish Intake Manually",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}/recover": {
      "post": {
        "operationId": "recover_intake_session_api_intake_sessions__session_id__recover_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recover Intake Session",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}/resume": {
      "post": {
        "operationId": "resume_intake_session_api_intake_sessions__session_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Intake Session",
        "tags": [
          "intake"
        ]
      }
    },
    "/api/intake/sessions/{session_id}/voice-token": {
      "post": {
        "operationId": "issue_browser_voice_token_api_intake_sessions__session_id__voice_token_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserVoiceTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Browser Voice Token",
        "tags": [
          "intake"
        ]
      }
    },
...
    "/api/jobs": {
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSpecV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      },
      "put": {
        "operationId": "replace_job_spec_api_jobs__job_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSpecV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Job Spec",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}