# services/api/app/api/router.py
"""Typed FastAPI routes for the mock-first VeraMove workflow."""

import json
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Depends, Header, Path, Query, Request, Response, status
from pydantic import ValidationError

from services.api.app.api.dependencies import (
    get_browser_voice_token_issuer,
    get_intake_recovery_service,
    get_intake_session_service,
    get_integration_status,
    get_live_voice_operator_service,
    get_service,
    get_settings,
    get_vendor_research_service,
)
from services.api.app.api.integration_status import IntegrationStatusSnapshot
from services.api.app.api.models import (
    AttachIntakeConversationRequest,
    BrowserVoiceTokenResponse,
    CreateIntakeSessionRequest,
    DocumentIntakeRequest,
    ElevenLabsConversationInitiationRequest,
    ElevenLabsConversationInitiationResponse,
    IntakeDynamicVariables,
    IntakeSessionResponse,
    JobEventsResponse,
    RuntimeHealthResponse,
)
from services.api.app.contracts import (
    HealthResponse,
    JobRecord,
    JobSpecV1,
    JobVendorResearchViewV1,
    RecommendationV1,
    VendorCallAuthorizationRequest,
    VendorDiscoveryResponse,
    VendorShortlistRequest,
    WebhookAck,
)
from services.api.app.core.config import Settings
from services.api.app.core.errors import ProviderRequestError, WebhookPayloadError
from services.api.app.integrations.elevenlabs.tokens import BrowserVoiceTokenIssuer
from services.api.app.orchestration.intake_recovery import IntakeRecoveryService
from services.api.app.orchestration.intake_sessions import (
    IntakeSession,
    IntakeSessionService,
    verify_pre_call_secret,
)
from services.api.app.orchestration.live_voice_operator import (
    LiveVoiceOperatorService,
)
from services.api.app.orchestration.service import VeraMoveService
from services.api.app.orchestration.vendor_research import VendorResearchService

router = APIRouter()
Service = Annotated[VeraMoveService, Depends(get_service)]
VendorResearch = Annotated[
    VendorResearchService,
    Depends(get_vendor_research_service),
]
RuntimeSettings = Annotated[Settings, Depends(get_settings)]
IntakeSessions = Annotated[IntakeSessionService, Depends(get_intake_session_service)]
IntakeRecovery = Annotated[
    IntakeRecoveryService,
    Depends(get_intake_recovery_service),
]
BrowserVoiceTokens = Annotated[
    BrowserVoiceTokenIssuer,
    Depends(get_browser_voice_token_issuer),
]
IntegrationStatus = Annotated[IntegrationStatusSnapshot, Depends(get_integration_status)]
LiveVoiceOperator = Annotated[
    LiveVoiceOperatorService,
    Depends(get_live_voice_operator_service),
]


def _intake_dynamic_variables(session: IntakeSession) -> IntakeDynamicVariables:
    base = session.base_job_spec
    return IntakeDynamicVariables(
        job_id=session.job_id,
        intake_session_id=session.intake_session_id,
        agent_config_version=session.agent_config_version,
        intake_data_mode=session.data_mode,
        resume_mode="structured_partial" if base is not None else "fresh",
        partial_job_spec_json=(
            base.model_dump_json(exclude_none=False) if base is not None else "{}"
        ),
        missing_fields_json=(
            json.dumps(base.missing_required_fields(), separators=(",", ":"))
            if base is not None
            else "[]"
        ),
    )


@router.get(
    "/health",
    response_model=HealthResponse | RuntimeHealthResponse,
    tags=["system"],
)
def health(settings: RuntimeSettings) -> RuntimeHealthResponse:
    return RuntimeHealthResponse(mode=settings.app_mode)


@router.get(
    "/api/integrations/status",
    response_model=IntegrationStatusSnapshot,
    tags=["system"],
)
def integration_status(snapshot: IntegrationStatus) -> IntegrationStatusSnapshot:
    return snapshot


@router.post(
    "/api/intake/document",
    response_model=JobRecord,
    status_code=status.HTTP_201_CREATED,
    tags=["intake"],
)
def create_job_from_document(
    request: DocumentIntakeRequest,
    service: Service,
) -> JobRecord:
    return service.create_job_from_document(request.document_text)


@router.post(
    "/api/intake/sessions",
    response_model=IntakeSessionResponse,
    status_code=status.HTTP_201_CREATED,
    tags=["intake"],
)
def create_intake_session(
    sessions: IntakeSessions,
    request: CreateIntakeSessionRequest | None = None,
) -> IntakeSessionResponse:
    data_mode = (
        request.data_mode
        if request is not None
        else CreateIntakeSessionRequest().data_mode
    )
    return IntakeSessionResponse.model_validate(
        sessions.create_web_session(data_mode=data_mode).model_dump()
    )


@router.get(
    "/api/intake/sessions/{session_id}",
    response_model=IntakeSessionResponse,
    tags=["intake"],
)
def get_intake_session(
    session_id: UUID,
    sessions: IntakeSessions,
) -> IntakeSessionResponse:
    return IntakeSessionResponse.model_validate(sessions.get_session(session_id).model_dump())


@router.post(
    "/api/intake/sessions/{session_id}/recover",
    response_model=IntakeSessionResponse,
    tags=["intake"],
)
def recover_intake_session(
    session_id: UUID,
    recovery: IntakeRecovery,
) -> IntakeSessionResponse:
    return IntakeSessionResponse.model_validate(
        recovery.recover(session_id).model_dump()
    )


@router.post(
    "/api/intake/sessions/{session_id}/resume",
    response_model=IntakeSessionResponse,
    tags=["intake"],
)
def resume_intake_session(
    session_id: UUID,
    recovery: IntakeRecovery,
) -> IntakeSessionResponse:
    return IntakeSessionResponse.model_validate(
        recovery.resume(session_id).model_dump()
    )


@router.post(
    "/api/intake/sessions/{session_id}/finish-manually",
    response_model=JobRecord,
    tags=["intake"],
)
def finish_intake_manually(
    session_id: UUID,
    recovery: IntakeRecovery,
) -> JobRecord:
    return recovery.finish_manually(session_id)


@router.get(
    "/api/intake/conversations/{conversation_id}",
    response_model=IntakeSessionResponse,
    tags=["intake"],
)
def get_intake_session_by_conversation(
    conversation_id: Annotated[str, Path(min_length=1, max_length=200)],
    sessions: IntakeSessions,
) -> IntakeSessionResponse:
    return IntakeSessionResponse.model_validate(
        sessions.get_by_conversation(conversation_id).model_dump()
    )


@router.post(
    "/api/intake/sessions/{session_id}/voice-token",
    response_model=BrowserVoiceTokenResponse,
    tags=["intake"],
)
def issue_browser_voice_token(
    session_id: UUID,
    response: Response,
    sessions: IntakeSessions,
    token_issuer: BrowserVoiceTokens,
) -> BrowserVoiceTokenResponse:
    session = sessions.reserve_browser_credential(session_id)
    try:
        token = token_issuer.issue_token()
    except ProviderRequestError:
        sessions.fail_session(session_id, "browser_token_issue_failed")
        raise
    response.headers["Cache-Control"] = "no-store"
    return BrowserVoiceTokenResponse(
        conversation_token=token,
        dynamic_variables=_intake_dynamic_variables(session),
    )


@router.post(
    "/api/intake/sessions/{session_id}/conversation",
    response_model=IntakeSessionResponse,
    tags=["intake"],
)
def attach_browser_voice_conversation(
    session_id: UUID,
    request: AttachIntakeConversationRequest,
    sessions: IntakeSessions,
) -> IntakeSessionResponse:
    sessions.attach_conversation(
        session_id,
        request.conversation_id,
        agent_id=sessions.expected_agent_id,
    )
    return IntakeSessionResponse.model_validate(
        sessions.get_session(session_id).model_dump()
    )


@router.post(
    "/api/webhooks/elevenlabs/pre-call",
    response_model=ElevenLabsConversationInitiationResponse,
    tags=["webhooks"],
    openapi_extra={
        "requestBody": {
            "required": True,
            "content": {
                "application/json": {
                    "schema": ElevenLabsConversationInitiationRequest.model_json_schema()
                }
            },
        }
    },
)
async def elevenlabs_conversation_initiation(
    request: Request,
    sessions: IntakeSessions,
    settings: RuntimeSettings,
) -> ElevenLabsConversationInitiationResponse:
    verify_pre_call_secret(
        settings.live_voice.precall_secret,
        request.headers.get("x-veramove-precall-secret"),
    )
    body = await request.body()
    if len(body) > 32_768:
        raise WebhookPayloadError("ElevenLabs conversation-initiation body is too large")
    try:
        raw_payload = json.loads(body)
        parsed = ElevenLabsConversationInitiationRequest.model_validate(raw_payload)
    except (json.JSONDecodeError, UnicodeDecodeError, ValidationError, TypeError) as exc:
        raise WebhookPayloadError("ElevenLabs conversation-initiation payload is invalid") from exc
    session = sessions.create_pre_call_session(
        agent_id=parsed.agent_id,
        provider_call_key=parsed.call_sid,
    )
    return ElevenLabsConversationInitiationResponse(
        dynamic_variables=IntakeDynamicVariables(
            job_id=session.job_id,
            intake_session_id=session.intake_session_id,
            agent_config_version=session.agent_config_version,
            intake_data_mode=session.data_mode,
        )
    )


@router.post(
    "/api/jobs",
    response_model=JobRecord,
    status_code=status.HTTP_201_CREATED,
    tags=["jobs"],
)
def create_job(job_spec: JobSpecV1, service: Service) -> JobRecord:
    return service.create_job(job_spec)


@router.get("/api/jobs/{job_id}", response_model=JobRecord, tags=["jobs"])
def get_job(job_id: UUID, service: Service) -> JobRecord:
    return service.get_job(job_id)


@router.put("/api/jobs/{job_id}", response_model=JobRecord, tags=["jobs"])
def replace_job_spec(
    job_id: UUID,
    job_spec: JobSpecV1,
    service: Service,
) -> JobRecord:
    return service.replace_job_spec(job_id, job_spec)


@router.get(
    "/api/jobs/{job_id}/events",
    response_model=JobEventsResponse,
    tags=["jobs"],
)
def get_job_events(job_id: UUID, service: Service) -> JobEventsResponse:
    return JobEventsResponse(events=service.get_events(job_id))


@router.post("/api/jobs/{job_id}/confirm", response_model=JobRecord, tags=["jobs"])
def confirm_job(job_id: UUID, service: Service) -> JobRecord:
    return service.confirm_job(job_id)


@router.get(
    "/api/jobs/{job_id}/vendor-research",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def get_vendor_research(
    job_id: UUID,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    return research.view(job_id)


@router.post(
    "/api/jobs/{job_id}/vendor-research/discover",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def discover_job_vendors(
    job_id: UUID,
    research: VendorResearch,
    refresh: Annotated[bool, Query()] = False,
) -> JobVendorResearchViewV1:
    research.discover(job_id, refresh=refresh)
    return research.view(job_id)


@router.put(
    "/api/jobs/{job_id}/vendor-research/shortlist",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def save_vendor_shortlist(
    job_id: UUID,
    request: VendorShortlistRequest,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    research.set_shortlist(job_id, request.vendor_ids)
    return research.view(job_id)


@router.delete(
    "/api/jobs/{job_id}/vendor-research/shortlist",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def clear_vendor_shortlist(
    job_id: UUID,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    research.clear_shortlist(job_id)
    return research.view(job_id)


@router.post(
    "/api/jobs/{job_id}/vendor-research/analyze",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def analyze_vendor_websites(
    job_id: UUID,
    research: VendorResearch,
    refresh: Annotated[bool, Query()] = False,
) -> JobVendorResearchViewV1:
    research.analyze(job_id, refresh=refresh)
    return research.view(job_id)


@router.post(
    "/api/jobs/{job_id}/vendor-research/contacts",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def extract_vendor_contacts(
    job_id: UUID,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    return research.extract_contacts(job_id)


@router.put(
    "/api/jobs/{job_id}/vendor-research/call-authorizations",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def authorize_vendor_calls(
    job_id: UUID,
    request: VendorCallAuthorizationRequest,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    return research.authorize_calls(job_id, request)


@router.delete(
    "/api/jobs/{job_id}/vendor-research/call-authorizations",
    response_model=JobVendorResearchViewV1,
    tags=["vendors"],
)
def clear_vendor_call_authorizations(
    job_id: UUID,
    research: VendorResearch,
) -> JobVendorResearchViewV1:
    return research.clear_call_authorizations(job_id)


@router.post("/api/jobs/{job_id}/calls", response_model=JobRecord, tags=["calls"])
def start_calls(job_id: UUID, service: Service) -> JobRecord:
    return service.start_calls(job_id)


@router.post("/api/jobs/{job_id}/negotiate", response_model=JobRecord, tags=["negotiation"])
def negotiate(job_id: UUID, service: Service) -> JobRecord:
    return service.negotiate(job_id)


@router.get("/api/jobs/{job_id}/report", response_model=RecommendationV1, tags=["reports"])
def get_report(job_id: UUID, service: Service) -> RecommendationV1:
    return service.get_report(job_id)


@router.get(
    "/api/calls/{call_id}/recording",
    response_class=Response,
    responses={
        200: {
            "content": {
                "audio/mpeg": {},
                "audio/mp4": {},
                "audio/wav": {},
            },
            "description": "Validated provider recording audio.",
        }
    },
    tags=["calls"],
)
def get_call_recording(
    call_id: UUID,
    job_id: Annotated[UUID, Query()],
    signature: Annotated[str, Query(pattern=r"^[a-f0-9]{64}$")],
    operator: LiveVoiceOperator,
) -> Response:
    payload = operator.fetch_recording(call_id, job_id, signature)
    return Response(
        content=payload.content,
        media_type=payload.media_type,
        headers={
            "Cache-Control": payload.cache_control,
            "Content-Length": str(payload.content_length),
        },
    )


@router.post(
    "/api/calls/{call_id}/repair",
    response_model=WebhookAck,
    tags=["calls"],
)
def repair_call(
    call_id: UUID,
    operator: LiveVoiceOperator,
    service: Service,
    operator_secret: Annotated[
        str | None,
        Header(alias="x-veramove-operator-secret", max_length=512),
    ] = None,
) -> WebhookAck:
    repair = operator.prepare_repair(call_id, operator_secret)
    return service.handle_elevenlabs_repair(repair)


@router.post(
    "/api/webhooks/elevenlabs",
    response_model=WebhookAck,
    tags=["webhooks"],
    openapi_extra={
        "requestBody": {
            "required": True,
            "content": {
                "application/json": {
                    "schema": {
                        "anyOf": [
                            {"$ref": "#/components/schemas/ElevenLabsWebhookEvent"},
                            {"$ref": "#/components/schemas/ElevenLabsPostCallWebhook"},
                        ]
                    },
                }
            },
        }
    },
)
async def elevenlabs_webhook(
    request: Request,
    service: Service,
) -> WebhookAck:
    return service.handle_elevenlabs_webhook(
        await request.body(),
        request.headers.get("elevenlabs-signature"),
    )


@router.get("/api/vendors/discover", response_model=VendorDiscoveryResponse, tags=["vendors"])
def discover_vendors(
    service: Service,
    origin: Annotated[str | None, Query(max_length=200)] = None,
    destination: Annotated[str | None, Query(max_length=200)] = None,
) -> VendorDiscoveryResponse:
    return VendorDiscoveryResponse(
        vendors=service.discover_vendors(origin, destination),
        source=service.vendor_discovery_source,
    )